requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
fastapi==0.65.1
h11==0.12.0
idna==2.10
pydantic==1.8.2
python-dotenv==0.17.1
requests==2.25.1
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
